+++ unpack.sh
@@ -1,5 +1,12 @@
 #!/bin/bash
 
+# Copyright © 2017–2018 University of Malta
+
+# Copying and distribution of this file, with or without modification,
+# are permitted in any medium without royalty provided the copyright
+# notice and this notice are preserved. This file is offered as-is,
+# without any warranty.
+
 # This script untars gmp, mpfr, mpc and tweaks them a little.
 
 set -e