+++ RELEASES.md
@@ -1,3 +1,10 @@
+<!-- Copyright © 2017–2018 University of Malta -->
+
+<!-- Copying and distribution of this file, with or without
+modification, are permitted in any medium without royalty provided the
+copyright notice and this notice are preserved. This file is offered
+as-is, without any warranty. -->
+
 Version 1.1.6 (2018-05-29)
 ==========================