+++ RELEASES.md
@@ -5,6 +5,11 @@ modification, are permitted in any medium without royalty provided the
 copyright notice and this notice are preserved. This file is offered
 as-is, without any warranty. -->
 
+Version 1.1.7 (unreleased)
+==========================
+
+* Update MPFR 4.0.1-p6 -> 4.0.1-p9.
+
 Version 1.1.6 (2018-05-29)
 ==========================